backend/services/7-forecasting-service/requirements.txt
fastapi
uvicorn[standard]
pandas
statsmodels
asyncpg
python-dotenv 